Issue number: Obtained from: Submitted by: Reviewed by:  Rajesh,Aravindan
Synopsis :Adding a sub-organistation using Access Manager 7.0 in JES4 breaks the gateway and Access Manager

Evaluation & Fix:
-------------------
Could reproduce this issue if the suborg which starts with the root suffix string is created.
As amadmin create root org : o=isp and a suborg as 'isp.abc.com',
Restart the server, login as amadmin again, getting the 'Contact your system admin..' error in the very first adminframe. Logged out.
Checked the debug logs. Noticed a 'StringIndexOutOfRange' error on a substring operation.
Fixed code to avoid that exception.


git-svn-id: https://svn.java.net/svn/opensso~svn/trunk@36 b2119f5d-607e-cfcf-c5d3-9fa3ed8061d4