Issue number: Obtained from: Submitted by: Reviewed by: Aravindan,Dennis,Rajesh Synopsis : non 'special' users are returned in call to getSpecialIdentities()
Evaluation & Fix:
-----------------
Based on the "SpecialUser plugin" SpecialRepo we are searching all the
supported identities from the idRepo Service and returning to console.
If the IdType is USER, we should remove "amadmin" and "anonymous" from
the list of special users.
Modified idm/AMIdentityRepository code to achieve this.
Added these two constants to IdConstants.java.
Modified AMIdentityRepository and SpecialRepo to remove these literals and use the constants from IdConstants.java.Added javadocs for the constants.


git-svn-id: https://svn.java.net/svn/opensso~svn/trunk@66 b2119f5d-607e-cfcf-c5d3-9fa3ed8061d4