Issue number: Obtained from: Submitted by: Reviewed by:  Aravindan,Dennis Description: ------------ Synopsis : Cannot delete an organization that got created under a container.
Evaluation & Fix:
---------------
In the Legacy mode/new console, under Directory Managment/Container, the
container gets created as ou=NewCont1,dc=iplanet,dc=com
The Organization under this Container gets created as
o=testContOrg,ou=NewCont1,dc=iplanet,dc=com

While converting the DN format to slash seperated format, the DNMapper
code parses it incorrectly as "testContOrgou=NewCont1" and it gets
displayed as a subrealm under the Access Control tab which is not correct.
Fixed the code to get the naming attribute for organization unit and
replace the strings accordingly so that it will result as organization
only under the Container.


git-svn-id: https://svn.java.net/svn/opensso~svn/trunk@90 b2119f5d-607e-cfcf-c5d3-9fa3ed8061d4