Issue number: Obtained from: Submitted by: Reviewed by: Bina,Aravindan *Synopsis*: realmMode attribute value not changed in Global Schema after migration to Realm Mode.
Modified the migration code to set the new value for this 'realm'
attribute in idrepo service's global schema.
Currently after migration, this new value is set only in the
configuration part of the idrepo service, but not in the Schema part.
This should be similar to the fresh realm only mode of installation.


git-svn-id: https://svn.java.net/svn/opensso~svn/trunk@91 b2119f5d-607e-cfcf-c5d3-9fa3ed8061d4