Issue number: Obtained from: Submitted by: Reviewed by: Aravindan,Rajesh Synopsis : build8/legacy:Get exception error when removing and adding session & Discovery service.
Evaluation & fix:
------------
While registering a service to a  realm/subrealm, we create the COS
templates as well the COS definition for dynamic service attribute
services.
The same way while unregistering the service from a realm/subrealm, we
need to delete the COS template and definition.
So the issue here is while registering the same service again, if logs
this Entry already exists error/exception for the COS definition entry
since the COS definition is not getting deleted.
In AMSDKRepo, we are deleting only the COS template and not the COS
definition.
Also we are not logging or throwing exception if the COS definition
entry exists. The decision made is to retain the COS definition for
those dynamic attributes services.
The fix is to change the debug.error to debug.warning in
ldap/DirectoryManager.java to ignore the existence of the COS definition
entry while registering the services (dynamic).


git-svn-id: https://svn.java.net/svn/opensso~svn/trunk@95 b2119f5d-607e-cfcf-c5d3-9fa3ed8061d4