Issue number: Obtained from: Submitted by: Reviewed by:  Kenneth,Rajesh *Synopsis*: Missing IdRepo isExist API in AMIdentity.java amadmin needs a AMIdentity.isExists method to determine the existence of an AMIdentity.
Evaluation & fix:
-----------------
1) Added this isExists() API to AMIdentity.java
Added code to call the AMDirectoryManager::isExists() which in turn calls the configured plugin's isExist() api to look for the particular AMIdentity object in that repository and returns a boolean value accordingly.

2) Fixed a typo 'flase' to 'false' in AMIdentity.java

3) Fixed a bug in idm/plugins/internal/SpecialRepo.java::isExists() method.
A third argument has to be passed to IdRepoUnsupportedOpException(..) as the key '305' and the value in amIdRepo.properties are looking for three parameters to be replaced.

305=Plug-in {0} does not support operation {1} for type {2}
SpecialRepo::isExists()
Object args[] = {NAME, IdOperation.SERVICE.getName()};

Otherwise it was displaying
com.sun.identity.idm.plugins.internal.SpecialRepo does not support operation
service for type {2}

4) Fixed sdk/AMDirectoryManager.java::isExists method to look for the hidden SpecialRepo plugin and not include that hidden repo to check if it exists or not. Added synchronized for idRepoMap to get the plugins which got added earlier, to ensure lock if other methods are updating that idRepoMap.


git-svn-id: https://svn.java.net/svn/opensso~svn/trunk@98 b2119f5d-607e-cfcf-c5d3-9fa3ed8061d4