Issue number: Obtained from: Submitted by: Reviewed by: Rajesh Synopsis: "Profile attributes, dn and entrydn, are not returned to the Agent with amSDK repo plugin"
The logic in this api is
1) to construct the com/iplanet/ums/PersistentObject based on the "DN" requested.
2) Get all the attribute names for that entry through "com/iplanet/ums/PersistentObject::getAttributeNames()" which does a LDAPConnection::read(String DN).
3) Get all the attribute values for the attribute names obtained for that particular requested "DN" through
"com/iplanet/ums/PersistentObject::getAttributes(for attributeNames obtained earlier)".

The search happens correctly with the query "attrs=ALL". (entry from DS access log pasted below).
The search result for getAttributeNames has all the attribute names for that requested "DN".
ldapjdk(LDAPConnection::read(dn)) does not include the requested "DN" in the result, but has all the attribute names for that entry returned. So that's the reason we are not seeing "dn=[...]" in the result from the call AMSDKRepo::getAttributes(..). This is only for AMSDK repo.
It works fine for LDAPV3 repo & Special Repo because the entrydn/dn is queried from the data store configuration and it is configured as part of user attributes in the data store configuration like
<AttributeSchema name="sun-idrepo-ldapv3-config-user-attributes"
type="list"
syntax="string"
i18nKey="a20171">
<DefaultValues>
<Value>cn</Value>
<Value>dn</Value>

Fix:
-----
To be consistent with other datastore getAttributes(),
1) modified com/iplanet/am/sdk/ldap/DirectoryServicesImpl::getAttributes().Gets all attributes corresponding to the entryDN.
1a) After obtaining the result from ldap, added the entry "dn" with the value of the "entryDN" passed to that result.

After the fix, verified that getAttributes() is returning all attributes including the entryDN for that entry.


git-svn-id: https://svn.java.net/svn/opensso~svn/trunk@139 b2119f5d-607e-cfcf-c5d3-9fa3ed8061d4