Issue number: 334 Obtained from: Submitted by: Reviewed by:  Aravindan,Dennis 1) Removed the code (SMSFlatFileObjectBase.java) where it reads the root suffix from "com.iplanet.am.defaultOrg" in AMConfig.properties.
2) Since FlatFile and ldap framework already extends SMSObjectDB, removed the
original code from SMSLdapObject.java::initialize() and introduced the
implementation in SMSObjectDB.java to get the base dn from serverconfig.xml's
'sms' server group and modified SMSFlatFileObjectBase.java to call the
getRootSuffix() from the parent.
3) Modified SMSEntry to handle the flatfile classname when the SMSObject is null
and not instantiate the default SMSLdapObject which was attempting to connect to
ldap server in the case of flatfile.


git-svn-id: https://svn.java.net/svn/opensso~svn/trunk@991 b2119f5d-607e-cfcf-c5d3-9fa3ed8061d4