Issue number: 385 Obtained from: Submitted by: Reviewed by:  Dennis,Rajesh,Aravindan,Ken Fix for amsdk doesn't fallback to primary directory server once primary comes up. 10) LDAPConnPoolUtils.java Placeholder for the ConnectionPool status. 1) LDAPConnectionPool.java code: While releasing the connection, in LDAPConnectionPool::close(), checked the LDAPException retry error codes (80-OTHER,81-SERVER_DOWN,91-CONNECT_ERROR) and initiated the failover instead of utilising the ldap jdk's failover mechanism. Marked those servers as 'down' when the connection got released. Also invoked the FallBackManager thread here which looks for the connection for every few minutes which was configured in AMConfig.properties  "com.sun.am.ldap.fallback.sleep.minutes" and pings if the primary is up for servers which were marked 'down' earlier while releasing the connection. If it is up and running falls back to primary server.
2) FallBackManager.java code:
Gets the time interval from AMConfig.properties set into the key
"com.sun.am.ldap.fallback.sleep.minutes" and
sets the time interval to check whether primary is up to fallback to
happen. If it is not set in the properties file,
default value of 15 minutes is set.
When the primary server is up, marks them as up and running for future reference.
Also invokes the EventService::resetAllSearches to facilitate event service/persistant
notifications to fallback.


git-svn-id: https://svn.java.net/svn/opensso~svn/trunk@1115 b2119f5d-607e-cfcf-c5d3-9fa3ed8061d4