Issue number: 409 Obtained from: Submitted by: Reviewed by:  Kenneth Lui We noticed these issues while testing opensso j2ee agent with AM 7.0 and AM 7.0 + patch. 1) Incorrect universal id, had amsdkdn embedded in it. Fix : Change in com/sun/identity/idm/AMIdentity.java
2) Newly added api in opensso client didn't have corresponding api on the old AM server.
Fix : Code changes for the backward compatibility (getAMSdkBaseDN)

3) Noticed ClassNotFound - DelegationEvaluator due to the permission check made for both client as well server.
Fix: Since it is a JAXRPC call, the permission checking would be done at the server. Client need not have this check. So change in com/sun/identity/sm/OrgConfigViaAMSDK made to avoid
checking the realm permission in case of SMSJAXRPCObject/client
4) 500 error thrown by server when notification url is not being
constructed properly on the client side.


git-svn-id: https://svn.java.net/svn/opensso~svn/trunk@1131 b2119f5d-607e-cfcf-c5d3-9fa3ed8061d4