Issue number:  3815 Obtained from: Submitted by:  Heng-Ming Hsu Reviewed by:   Ping Luo
Client.java try to construct com.sun.net.ssl.internal.ssl.Provider and move it to fisrt entry of
security provider. However, WAS6.1 has its own jdk with doesn't have this provider. The fix
is to check if com.sun.net.ssl.internal.ssl.Provider is there. If not, it will not move it.


git-svn-id: https://svn.java.net/svn/opensso~svn/trunk@5070 b2119f5d-607e-cfcf-c5d3-9fa3ed8061d4