Issue number:  3474 Obtained from: Submitted by:  dennis Reviewed by:   kevin Bootstrap file cannot be sync up on GFv3 on Solaris It does not happen on all machines, Nicolas' setup (Web Server 7 u3 on solaris) has this problem too.
Basically, after configuration and when web server starts
up, Bootstrap file is being sync-ed up with serverconfig.xml
which is in the server.

Since bootstrap file is readonly, we programmicatically
check it to writable and write to it. However, there is
a time delay for the System to make it readable and hence
writing to the file


git-svn-id: https://svn.java.net/svn/opensso~svn/trunk@5387 b2119f5d-607e-cfcf-c5d3-9fa3ed8061d4