Issue number:  4331 Obtained from: Submitted by:  dennis Reviewed by:   qingwen Step to reproduce. 1. use ssoadm create-metadata-templ sub command to create metadata and extended metadata template 2. edit the extended metadata template to include a COT (let it be "test") 3. use ssoadm import-entity sub command to import both XML with --cot test 4. got an error despite of success operation
Unable to add role to entity descriptor, xyz because it already existed.

Fix
this is because the entity is already added to the COT, test
and we attempt to add it again.
added code to not add it to COT if it is already added.

also find a more serious issue.
SystemProperties is not mapping property name to attribute name in service.
(this is done when we move properties from AMConfig.properties to
 individual component services)
 to resolve this, add a System property to indicate that ssoadm is running
 and find the correct the values in services accordingly.


git-svn-id: https://svn.java.net/svn/opensso~svn/trunk@5573 b2119f5d-607e-cfcf-c5d3-9fa3ed8061d4