Issue number:  5502 Obtained from: Submitted by:  dennis Reviewed by:   qingwen This issue is found in Express 8 and not in the TRUNK workspace. the reason is because the entitlement workspace is integrated in opensso/products after Express 8 RTM CVS branch is created.
the reason why it happens in Express 8 is because console classes
are find in two jars namely the amserver.jar and console.jar.
the console classes are identical in both jars in the TRUNK
(hence this bug does not exist); however they are different
(specifically, com.sun.identity.console.policy.model.PolicyModelImpl.class)
in both jars in Express 8 branch.

the web container used the PolicyModelImpl.class in amserver.jar
and drop the one in console.jar. and resulted in policy cannot be
created in sub realm.

the workaround for Express 8 is to use CLI or ssoadm.jsp to create policies.

and we have to make sure that amserver.jar does not contain console class in
the TRUNK


git-svn-id: https://svn.java.net/svn/opensso~svn/trunk@7827 b2119f5d-607e-cfcf-c5d3-9fa3ed8061d4