Issue number:  5516 Obtained from: Submitted by:  dennis Reviewed by:   aravindan case 1: console is not showing the org aliases in the rule page this is because console is not calling the correct API to determine if the realm and service type can create new resources or not. It is calling the can create policy API instead of can create resources API
case 2: server is in a unusable state.
this is because of indefinite looping.
During policy evaluation, referral policies are created
when "Realm Alias Referrals" option is on.
However, referral policy is also created for root realm
when we only expected referral policies to be created only
for sub realms. Since root referral policy is created,
the policy evaluation goes into a loop.


git-svn-id: https://svn.java.net/svn/opensso~svn/trunk@7866 b2119f5d-607e-cfcf-c5d3-9fa3ed8061d4