Issue number:  6087 Obtained from: Submitted by:  dennis Reviewed by:   aravindan issue 1 The issue is the policy admin does not have the permissions to remove entry in the sunEntitlementIndexes node in SM datastore. Hence policy entry under remove in the iPlanetAMPolicyService node but not under the sunEntitlementIndexes node
To fix this issue, switch to use super admin credential to remove
entry under sunEntitlementIndexes node.

issue 2
after the referral policy is deleted in root realm, we are trying
to clear the index cache in the sub realm after deleting the policy
in sub realm. We are unable to get the application object in the sub realm
to construct the resource save index since referral policy has already
deleted.

to fix this issue, we get the application object in root realm if
we are unable to get it in sub realm


git-svn-id: https://svn.java.net/svn/opensso~svn/trunk@8592 b2119f5d-607e-cfcf-c5d3-9fa3ed8061d4